# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "restaurantAPI",
        "version": "0.1"
    },
    "paths": {
...
        "/api/users": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Вывод всех пользователей",
                "operationId": "usersAll",
                "responses": {
                    "200": {
                        "description": "Пользователи выведены"
                    },
                    "403": {
                        "description": "Нет доступа или пользователь не авторизован"
                    }
                }
            },
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Создание пользователя",
                "operationId": "userCreate",
                "responses": {
                    "200": {
                        "description": "Пользователь создан"
                    },
                    "403": {
                        "description": "Нет доступа или пользователь не авторизован"
                    }
                }
            }
        },
        "/api/users/{user_id}": {
            "put": {
                "tags": [
                    "Users"
                ],
                "summary": "Изменение пользователя по id",
                "operationId": "userUpdate",
                "responses": {
                    "200": {
                        "description": "Пользователь изменен"
                    },
                    "403": {
                        "description": "Нет доступа или пользователь не авторизован"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Users"
                ],
                "summary": "Удаление пользователя по id",
                "operationId": "userDelete",
                "responses": {
                    "200": {
                        "description": "Пользователь удален"
                    },
                    "403": {
                        "description": "Нет доступа или пользователь не авторизован"
                    }
                }
            }
        }
    }
}